requirements.txt
sqlalchemy
pandas
requests
boto3
psycopg2-binary
python-dotenv
pymysql
pyodbc